Refactor: Seperation hooks and component
