+++ src/hooks/useUploder.ts
@@ -1,0 +1,38 @@
+import { useState } from 'react';
+import type { UploadFile, UploadProps } from 'antd';
+import type { GetProp } from 'antd';
+
+type FileType = Parameters<GetProp<UploadProps, 'beforeUpload'>>[0];
+
+const useUploader = () => {
+  const [fileList, setFileList] = useState<UploadFile[]>([]);
+  const [uploading, setUploading] = useState(false);
+
+  const handleUpload = () => {
+    const formData = new FormData();
+    fileList.forEach((file) => {
+      formData.append('files[]', file as FileType);
+    });
+    setUploading(true);
+    // 파일 업로드하는 로직 넣어야해요
+    //https://ant.design/components/upload 여기에서 Upload manually 참고
+  };
+
+  const uploadProps: UploadProps = {
+    onRemove: (file) => {
+      const index = fileList.indexOf(file);
+      const newFileList = fileList.slice();
+      newFileList.splice(index, 1);
+      setFileList(newFileList);
+    },
+    beforeUpload: (file) => {
+      setFileList([...fileList, file]);
+      return false;
+    },
+    fileList,
+  };
+
+  return { fileList, uploading, handleUpload, uploadProps };
+};
+
+export default useUploader;
